+++ src/app/components/Navbar.tsx
@@ -2,6 +2,9 @@ import * as React from 'react';
 import {useTokenState, useTokenDispatch} from '../store/TokenContext';
 import Icon from './Icon';
 import Tooltip from './Tooltip';
+import {pullRemoteTokens} from '../store/remoteTokens';
+import TokenData from './TokenData';
+import {StorageProviderType} from '../../types/api';
 
 const TabButton = ({name, label, active, setActive, first = false}) => (
     <button
@@ -17,8 +20,16 @@ const TabButton = ({name, label, active, setActive, first = false}) => (
 );
 
 const Navbar = ({active, setActive}) => {
-    const {colorMode} = useTokenState();
-    const {toggleColorMode, pullStyles} = useTokenDispatch();
+    const {colorMode, storageType, api} = useTokenState();
+    const {toggleColorMode, pullStyles, setLoading, setTokenData} = useTokenDispatch();
+
+    const handlePull = async () => {
+        setLoading(true);
+        const updatedTokens = await pullRemoteTokens(api);
+        setTokenData(new TokenData(updatedTokens), updatedTokens.updatedAt);
+        setLoading(false);
+    };
+
     return (
         <div className="sticky top-0 navbar bg-white flex items-center justify-between z-1 border-b border-gray-200">
             <div>
@@ -28,6 +39,13 @@ const Navbar = ({active, setActive}) => {
                 <TabButton name="settings" label="Settings" active={active} setActive={setActive} />
             </div>
             <div>
+                {storageType.provider !== StorageProviderType.LOCAL && (
+                    <Tooltip label={`Pull from ${storageType.provider}`}>
+                        <button onClick={handlePull} type="button" className="button button-ghost">
+                            <Icon name="refresh" />
+                        </button>
+                    </Tooltip>
+                )}
                 <Tooltip variant="right" label="Import Styles">
                     <button className="button button-ghost" type="button" onClick={pullStyles}>
                         <Icon name="import" />
