+++ src/app/components/icons/IconRefresh.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 
 const IconRefresh = () => (
-    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
         <path
             fillRule="evenodd"
             clipRule="evenodd"
-            d="M19.4264 9.54425L18.1025 11.1991C18.0958 11.1528 18.0886 11.1065 18.0809 11.0604C17.9236 10.1202 17.5459 9.22553 16.9728 8.45212C16.1992 7.40817 15.1105 6.6405 13.8674 6.26234C12.6243 5.88418 11.2926 5.91557 10.0687 6.35186C9.16194 6.67509 8.34996 7.20788 7.69578 7.90116C7.46695 8.14367 7.25743 8.40581 7.06991 8.6855L7.1547 8.74234L7.90051 9.24237L7.98529 9.29921C8.57728 8.41624 9.43739 7.74691 10.4387 7.38995C11.4401 7.033 12.5296 7.00732 13.5467 7.31671C14.5637 7.6261 15.4544 8.25417 16.0873 9.10828C16.6346 9.84675 16.9641 10.7204 17.0433 11.6305L14.5847 9.99151L14.03 10.8236L17.3357 13.0273L17.7171 13.2816L18.0035 12.9237L20.2073 10.1689L19.4264 9.54425ZM6.20377 11.1971L4 13.9518L4.78087 14.5765L6.10476 12.9216C6.11142 12.968 6.11861 13.0142 6.12633 13.0604C6.28362 14.0005 6.66135 14.8952 7.23448 15.6686C8.00809 16.7126 9.09674 17.4802 10.3398 17.8584C11.5829 18.2365 12.9147 18.2052 14.1386 17.7689C15.0453 17.4456 15.8573 16.9128 16.5115 16.2196C16.7403 15.9771 16.9498 15.7149 17.1373 15.4352L17.0526 15.3784L16.3067 14.8784L16.222 14.8215L16.2205 14.8236C15.6286 15.7056 14.7691 16.3741 13.7685 16.7308C12.7672 17.0877 11.6776 17.1134 10.6606 16.804C9.64354 16.4946 8.75286 15.8666 8.11993 15.0124C7.57269 14.274 7.24319 13.4003 7.164 12.4902L9.62251 14.1292L10.1772 13.2972L6.87155 11.0934L6.49013 10.8391L6.20377 11.1971Z"
+            d="M14.6096 5.68765L13.4349 7.15603C13.4268 7.10387 13.418 7.05185 13.4084 7C13.2569 6.18064 12.9203 5.40189 12.419 4.72539C11.7169 3.77797 10.7289 3.08128 9.60075 2.73808C8.47259 2.39489 7.264 2.42337 6.15326 2.81933C5.36015 3.10206 4.64691 3.56145 4.06479 4.15764C3.83166 4.3964 3.61956 4.6571 3.43178 4.93718L3.43192 4.93728L4.26237 5.49406L4.26252 5.49416C4.79977 4.69282 5.58035 4.08538 6.4891 3.76143C7.39785 3.43748 8.38666 3.41418 9.30967 3.69496C10.2327 3.97574 11.041 4.54574 11.6154 5.32088C12.1002 5.97502 12.3966 6.74603 12.4774 7.55063L10.2774 6.08398L9.7227 6.91603L12.7227 8.91603L13.1041 9.1703L13.3905 8.81235L15.3905 6.31235L14.6096 5.68765ZM2.60962 7.18765L0.609619 9.68765L1.39049 10.3123L2.56519 8.84397C2.57329 8.89614 2.58213 8.94815 2.59172 9C2.74323 9.81936 3.07981 10.5981 3.58113 11.2746C4.2832 12.222 5.27119 12.9187 6.39935 13.2619C7.52751 13.6051 8.7361 13.5766 9.84684 13.1807C10.64 12.8979 11.3532 12.4386 11.9353 11.8424C12.168 11.6041 12.3797 11.344 12.5672 11.0646L12.5683 11.0628L12.5682 11.0627L11.7377 10.5059L11.7376 10.5058L11.7365 10.5074C11.1993 11.308 10.4192 11.9148 9.51101 12.2386C8.60225 12.5625 7.61344 12.5858 6.69044 12.305C5.76744 12.0243 4.95911 11.4543 4.38471 10.6791C3.89996 10.025 3.60346 9.25397 3.52271 8.44937L5.7227 9.91603L6.2774 9.08398L3.2774 7.08398L2.89598 6.8297L2.60962 7.18765Z"
             fill="black"
             fillOpacity="0.8"
         />
